added the check mark and delete button to the add-to-do.js
