+++ client/src/init/index.js
@@ -1,1 +1,3 @@
-import '../listeners/listener.js';
+/* --- import all listener files so they can attach to the DOM --- */
+
+import '../listeners/add-to-do-listener.js';
