+++ client/src/listeners/add-to-do-listener.js
@@ -1,4 +1,4 @@
-import { addTodo } from '../handlers/handler.js';
+import { addTodo } from '../handlers/add-to-do.js';
 
 const todoButton = document.querySelector('.todo-button');
 // event listeners
