+++ docs/dependency-graph.html
@@ -53,25 +53,35 @@
 <!-- Generated by graphviz version 2.44.0 (20200408.0750)
  -->
 <!-- Title: dependency&#45;cruiser output Pages: 1 -->
-<svg width="126pt" height="173pt"
- viewBox="0.00 0.00 126.00 173.00" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
+<svg width="333pt" height="173pt"
+ viewBox="0.00 0.00 333.02 173.00" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">
 <g id="graph0" class="graph" transform="scale(1 1) rotate(0) translate(4 169)">
 <title>dependency&#45;cruiser output</title>
-<polygon fill="white" stroke="transparent" points="-4,4 -4,-169 122,-169 122,4 -4,4"/>
+<polygon fill="white" stroke="transparent" points="-4,4 -4,-169 329.02,-169 329.02,4 -4,4"/>
 <g id="clust1" class="cluster">
 <title>cluster_client</title>
-<path fill="#ffffff" stroke="black" stroke-width="2" d="M20,-8C20,-8 98,-8 98,-8 104,-8 110,-14 110,-20 110,-20 110,-145 110,-145 110,-151 104,-157 98,-157 98,-157 20,-157 20,-157 14,-157 8,-151 8,-145 8,-145 8,-20 8,-20 8,-14 14,-8 20,-8"/>
-<text text-anchor="middle" x="59" y="-144.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">client</text>
+<path fill="#ffffff" stroke="black" stroke-width="2" d="M20,-8C20,-8 305.02,-8 305.02,-8 311.02,-8 317.02,-14 317.02,-20 317.02,-20 317.02,-145 317.02,-145 317.02,-151 311.02,-157 305.02,-157 305.02,-157 20,-157 20,-157 14,-157 8,-151 8,-145 8,-145 8,-20 8,-20 8,-14 14,-8 20,-8"/>
+<text text-anchor="middle" x="162.51" y="-144.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">client</text>
 </g>
 <g id="clust2" class="cluster">
 <title>cluster_client/src</title>
-<path fill="#ffffff" stroke="black" stroke-width="2" d="M28,-42C28,-42 90,-42 90,-42 96,-42 102,-48 102,-54 102,-54 102,-118 102,-118 102,-124 96,-130 90,-130 90,-130 28,-130 28,-130 22,-130 16,-124 16,-118 16,-118 16,-54 16,-54 16,-48 22,-42 28,-42"/>
-<text text-anchor="middle" x="59" y="-117.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">src</text>
+<path fill="#ffffff" stroke="black" stroke-width="2" d="M28,-42C28,-42 297.02,-42 297.02,-42 303.02,-42 309.02,-48 309.02,-54 309.02,-54 309.02,-118 309.02,-118 309.02,-124 303.02,-130 297.02,-130 297.02,-130 28,-130 28,-130 22,-130 16,-124 16,-118 16,-118 16,-54 16,-54 16,-48 22,-42 28,-42"/>
+<text text-anchor="middle" x="162.51" y="-117.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">src</text>
 </g>
 <g id="clust3" class="cluster">
+<title>cluster_client/src/handlers</title>
+<path fill="#ffffff" stroke="black" stroke-width="2" d="M233.51,-50C233.51,-50 289.02,-50 289.02,-50 295.02,-50 301.02,-56 301.02,-62 301.02,-62 301.02,-91 301.02,-91 301.02,-97 295.02,-103 289.02,-103 289.02,-103 233.51,-103 233.51,-103 227.51,-103 221.51,-97 221.51,-91 221.51,-91 221.51,-62 221.51,-62 221.51,-56 227.51,-50 233.51,-50"/>
+<text text-anchor="middle" x="261.26" y="-90.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">handlers</text>
+</g>
+<g id="clust4" class="cluster">
 <title>cluster_client/src/init</title>
 <path fill="#ffffff" stroke="black" stroke-width="2" d="M36,-50C36,-50 82,-50 82,-50 88,-50 94,-56 94,-62 94,-62 94,-91 94,-91 94,-97 88,-103 82,-103 82,-103 36,-103 36,-103 30,-103 24,-97 24,-91 24,-91 24,-62 24,-62 24,-56 30,-50 36,-50"/>
 <text text-anchor="middle" x="59" y="-90.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">init</text>
+</g>
+<g id="clust5" class="cluster">
+<title>cluster_client/src/listeners</title>
+<path fill="#ffffff" stroke="black" stroke-width="2" d="M114,-50C114,-50 201.51,-50 201.51,-50 207.51,-50 213.51,-56 213.51,-62 213.51,-62 213.51,-91 213.51,-91 213.51,-97 207.51,-103 201.51,-103 201.51,-103 114,-103 114,-103 108,-103 102,-97 102,-91 102,-91 102,-62 102,-62 102,-56 108,-50 114,-50"/>
+<text text-anchor="middle" x="157.75" y="-90.9" font-family="Helvetica,sans-Serif" font-weight="bold" font-size="9.00">listeners</text>
 </g>
 <!-- client/data.js -->
 <g id="node1" class="node">
@@ -82,14 +92,44 @@
 </a>
 </g>
 </g>
-<!-- client/src/init/index.js -->
+<!-- client/src/handlers/add&#45;to&#45;do.js -->
 <g id="node2" class="node">
+<title>client/src/handlers/add&#45;to&#45;do.js</title>
+<g id="a_node2"><a xlink:href="../client/src/handlers/add-to-do.js" xlink:title="add&#45;to&#45;do.js">
+<path fill="#ffffcc" stroke="black" d="M286.57,-76.3C286.57,-76.3 235.95,-76.3 235.95,-76.3 232.85,-76.3 229.75,-73.2 229.75,-70.1 229.75,-70.1 229.75,-63.9 229.75,-63.9 229.75,-60.8 232.85,-57.7 235.95,-57.7 235.95,-57.7 286.57,-57.7 286.57,-57.7 289.67,-57.7 292.77,-60.8 292.77,-63.9 292.77,-63.9 292.77,-70.1 292.77,-70.1 292.77,-73.2 289.67,-76.3 286.57,-76.3"/>
+<text text-anchor="middle" x="261.26" y="-64.3" font-family="Helvetica,sans-Serif" font-size="9.00">add&#45;to&#45;do.js</text>
+</a>
+</g>
+</g>
+<!-- client/src/init/index.js -->
+<g id="node3" class="node">
 <title>client/src/init/index.js</title>
-<g id="a_node2"><a xlink:href="../client/src/init/index.js" xlink:title="index.js">
-<path fill="#ccffcc" stroke="black" d="M79.8,-76.3C79.8,-76.3 38.2,-76.3 38.2,-76.3 35.1,-76.3 32,-73.2 32,-70.1 32,-70.1 32,-63.9 32,-63.9 32,-60.8 35.1,-57.7 38.2,-57.7 38.2,-57.7 79.8,-57.7 79.8,-57.7 82.9,-57.7 86,-60.8 86,-63.9 86,-63.9 86,-70.1 86,-70.1 86,-73.2 82.9,-76.3 79.8,-76.3"/>
+<g id="a_node3"><a xlink:href="../client/src/init/index.js" xlink:title="index.js">
+<path fill="#ffffcc" stroke="black" d="M79.8,-76.3C79.8,-76.3 38.2,-76.3 38.2,-76.3 35.1,-76.3 32,-73.2 32,-70.1 32,-70.1 32,-63.9 32,-63.9 32,-60.8 35.1,-57.7 38.2,-57.7 38.2,-57.7 79.8,-57.7 79.8,-57.7 82.9,-57.7 86,-60.8 86,-63.9 86,-63.9 86,-70.1 86,-70.1 86,-73.2 82.9,-76.3 79.8,-76.3"/>
 <text text-anchor="middle" x="59" y="-64.3" font-family="Helvetica,sans-Serif" font-size="9.00">index.js</text>
 </a>
 </g>
+</g>
+<!-- client/src/listeners/add&#45;to&#45;do&#45;listener.js -->
+<g id="node4" class="node">
+<title>client/src/listeners/add&#45;to&#45;do&#45;listener.js</title>
+<g id="a_node4"><a xlink:href="../client/src/listeners/add-to-do-listener.js" xlink:title="add&#45;to&#45;do&#45;listener.js">
+<path fill="#ffffcc" stroke="black" d="M199.06,-76.3C199.06,-76.3 116.45,-76.3 116.45,-76.3 113.35,-76.3 110.25,-73.2 110.25,-70.1 110.25,-70.1 110.25,-63.9 110.25,-63.9 110.25,-60.8 113.35,-57.7 116.45,-57.7 116.45,-57.7 199.06,-57.7 199.06,-57.7 202.16,-57.7 205.26,-60.8 205.26,-63.9 205.26,-63.9 205.26,-70.1 205.26,-70.1 205.26,-73.2 202.16,-76.3 199.06,-76.3"/>
+<text text-anchor="middle" x="157.75" y="-64.3" font-family="Helvetica,sans-Serif" font-size="9.00">add&#45;to&#45;do&#45;listener.js</text>
+</a>
+</g>
+</g>
+<!-- client/src/init/index.js&#45;&gt;client/src/listeners/add&#45;to&#45;do&#45;listener.js -->
+<g id="edge1" class="edge">
+<title>client/src/init/index.js&#45;&gt;client/src/listeners/add&#45;to&#45;do&#45;listener.js</title>
+<path fill="none" stroke="#000000" stroke-width="2" stroke-opacity="0.200000" d="M86.18,-67C91.72,-67 97.76,-67 103.89,-67"/>
+<polygon fill="#000000" fill-opacity="0.200000" stroke="#000000" stroke-width="2" stroke-opacity="0.200000" points="104.04,-69.1 110.04,-67 104.04,-64.9 104.04,-69.1"/>
+</g>
+<!-- client/src/listeners/add&#45;to&#45;do&#45;listener.js&#45;&gt;client/src/handlers/add&#45;to&#45;do.js -->
+<g id="edge2" class="edge">
+<title>client/src/listeners/add&#45;to&#45;do&#45;listener.js&#45;&gt;client/src/handlers/add&#45;to&#45;do.js</title>
+<path fill="none" stroke="#000000" stroke-width="2" stroke-opacity="0.200000" d="M205.35,-67C211.4,-67 217.52,-67 223.38,-67"/>
+<polygon fill="#000000" fill-opacity="0.200000" stroke="#000000" stroke-width="2" stroke-opacity="0.200000" points="223.62,-69.1 229.62,-67 223.62,-64.9 223.62,-69.1"/>
 </g>
 </g>
 </svg>
